add image product

+++ src/main/java/com/example/productmanager/dto/request/ProductCreationRequest.java
@@ -15,5 +15,5 @@ public class ProductCreationRequest {
     String productDescription;
     String productImage;
     int productQuantity;
-    BigDecimal productPrice;
+    long productPrice;
 }
